put in error checking in the callstate handlers.  If a XMPP response had an error in handling of a call, then assume dropped call for now.

+++ OnSipCommon/StateMachine.h
@@ -135,8 +135,8 @@ public:
 	}
 
 	// # of msecs since the last time there was a state change
-	DWORD MsecsSinceLastStateChange()
-	{	return GetTickCount() - m_lastStateChange;	}
+	long MsecsSinceLastStateChange()
+	{	return (long) (GetTickCount() - m_lastStateChange);	}
 
 	Tstate getState()
 	{	return m_state; }
